rename project (#7)
