+++ .github/workflows/release.yml
@@ -30,5 +30,5 @@ jobs:
           platforms: linux/amd64,linux/arm64
           push: true
           tags: |
-            ghcr.io/aleksey925/transmission-telegram-bot:latest
-            ghcr.io/aleksey925/transmission-telegram-bot:${{ env.RELEASE_VERSION }}
+            ghcr.io/aleksey925/tg-trnsm-bot:latest
+            ghcr.io/aleksey925/tg-trnsm-bot:${{ env.RELEASE_VERSION }}
