+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "transmission-telegram-bot"
+name = "tg-trnsm-bot"
 version = "0.1.0"
 requires-python = ">=3.14, <4"
 dependencies = [
