+++ tg_trnsm_bot/__main__.py
@@ -1,0 +1,4 @@
+from tg_trnsm_bot.app import run
+
+if __name__ == "__main__":
+    run()
