--- transmission_telegram_bot/__main__.py
@@ -1,4 +1,0 @@
-from transmission_telegram_bot.app import run
-
-if __name__ == "__main__":
-    run()
